Move notification tests to a separate directory
